refactor layeroptionsmodal putting the logic out into a separate view, so it can be reused, fix replace, reintroduce data sliders
