+++ src/core/views/layers/LayerOptionsCoreView.css
