+++ src/core/views/layers/LayerOptionsModalView.hbs
@@ -4,63 +4,6 @@
       <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
       <h4 class="modal-title" style="display:inline">{{displayName}}</h4>
     </div>
-    <div class="modal-body">
-      <form class="form-horizontal">
-        <div class="form-group">
-          <label class="col-sm-2 control-label">{{t 'Opacity'}}</label>
-          <div class="col-sm-10">
-            <input class="form-control opacity-slider">
-          </div>
-        </div>
-        {{#each options}}
-          <div class="form-group">
-            <label for="visualization-selector_{{this.id}}" class="col-sm-2 control-label">
-              <input type="radio" id="visualization-selector_{{this.id}}" name="visualization-selector" class="visualization-selector" {{#if this.isChosen}} checked{{/if}}><span>{{t this.name}}</span>
-              </input>
-            </label>
-            {{#if this.selectThree}}
-              <div class="col-sm-4 option-row">
-                <select class="form-control layer-option" name="{{this.target}}">
-                  {{#each this.values}}
-                    <option value="{{this.value}}" {{#if this.isCurrentB1}} selected="selected"{{/if}}>{{this.label}}</option>
-                  {{/each}}
-                </select>
-              </div>
-              <div class="col-sm-4 option-row">
-                <select class="form-control layer-option" name="{{this.target}}">
-                  {{#each this.values}}
-                    <option value="{{this.value}}" {{#if this.isCurrentB2}} selected="selected"{{/if}}>{{this.label}}</option>
-                  {{/each}}
-                </select>
-              </div>
-              <div class="col-sm-4 option-row">
-                <select class="form-control layer-option" name="{{this.target}}">
-                  {{#each this.values}}
-                    <option value="{{this.value}}" {{#if this.isCurrentB3}} selected="selected"{{/if}}>{{this.label}}</option>
-                  {{/each}}
-                </select>
-              </div>
-            {{/if}}
-          {{#unless this.selectThree }}
-            <div class="col-sm-10 option-row">
-              <select class="form-control layer-option" name="{{this.target}}">
-                {{#each this.values}}
-                  <option value="{{this.value}}" {{#if this.isCurrentB1}} selected="selected"{{/if}}>{{this.label}}</option>
-                {{/each}}
-              </select>
-            </div>
-          {{/unless}}
-          </div>
-        {{/each}}
-        {{#if legendUrl}}
-          <div class="form-group">
-            <label class="col-sm-2 control-label">{{t 'Legend'}}</label>
-            <div class="col-sm-12">
-              <img class="layer-legend-image" src="{{legendUrl}}" alt="Legend" />
-            </div>
-          </div>
-        {{/if}}
-      </form>
-    </div>
+    <div class="modal-body"></div>
   </div>
 </div>
